Fix Spark 3.2.1 build failures and make it non-snapshot (#4508)
* Fix Spark 3.2.1 official release

* Fix up shim inheritance

* Fix 320  extends

* update buildall script

* Update copyrights


* update shuffle manager docs for 3.2.1

* fix missing ,

* minor changes in repos which will be removed later

* Update 33X shim

* remove temporary staging repositories

* update comment
